requirements.txt
accelerate==0.16.0
aiofiles==0.6.0
aiohttp==3.8.4
aiohttp-socks==0.7.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bitsandbytes==0.35.0
certifi==2022.12.7
charset-normalizer==3.0.1
docopt==0.6.2
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
googlevoice @ git+https://github.com/jaraco/googlevoice.git@17ba8a3b60588abd3af7c16c3e804dfe5c6ddca5
h11==0.12.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.12.1
hyperframe==6.0.1
idna==3.4
imap-tools==1.0.0
jsonschema==4.17.3
Logbook==1.5.3
matrix-client==0.4.0
matrix-nio==0.20.1
multidict==6.0.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pipreqs==0.4.11
psutil==5.9.4
pycryptodome==3.17
pyrsistent==0.19.3
python-socks==2.1.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
unpaddedbase64==2.1.0
urllib3==1.26.14
websockets==10.4
yarg==0.1.9
yarl==1.8.2
